[Refactor] AWS 재설정
