+++ src/main/java/com/sidenow/global/config/aws/service/AwsS3Service.java
@@ -1,4 +1,4 @@
-package com.sidenow.global.config.aws;
+package com.sidenow.global.config.aws.service;
 
 import com.sidenow.domain.boardType.free.board.entity.FreeBoard;
 import com.sidenow.domain.member.entity.Member;
